OT254-79/requirements.txt
greenlet==1.1.2
numpy==1.23.1
pandas==1.4.3
psycopg2-binary==2.9.3
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.39
